miniproject/trading/serializers.py
from rest_framework import serializers
from .models import Order
from .models import Transaction


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'
        read_only_fields = ['status', 'created_at', 'updated_at']



class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = '__all__'
        read_only_fields = ['executed_at']