miniproject/requirements.txt
Django~=5.1.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
celery
redis
psycopg2-binary
pillow
pdfkit
django-weasyprint
channels
